🚚 Move: `UserId` model to `crate::model::util::user_id` module [#93]
