+++ src/model/util/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod badge_id;
 pub mod timestamp;
+pub mod user_id;
